Refactor some onnx download utils
